+++ lib/hubbado/log.rb
@@ -1,5 +1,6 @@
 require 'dependency'
 require "hubbado/log/tags"
+require "hubbado/log/display"
 require "hubbado/log/configuration"
 require "hubbado/log/log"
 require "hubbado/log/logger"